Refactor maintenance services out of main
