+++ app/maintenance/__init__.py
@@ -1,0 +1,29 @@
+from .service import (
+    BACKUPS_DIR_REL,
+    REPLICATION_ALLOWED_PREFIXES,
+    REPLICATION_STATE_REL,
+    REPLICATION_TOMBSTONES_REL,
+    backup_create_service,
+    backup_restore_test_service,
+    compact_run_service,
+    iter_replication_files,
+    load_replication_state,
+    metrics_service,
+    replication_pull_service,
+    replication_push_service,
+)
+
+__all__ = [
+    "BACKUPS_DIR_REL",
+    "REPLICATION_ALLOWED_PREFIXES",
+    "REPLICATION_STATE_REL",
+    "REPLICATION_TOMBSTONES_REL",
+    "backup_create_service",
+    "backup_restore_test_service",
+    "compact_run_service",
+    "iter_replication_files",
+    "load_replication_state",
+    "metrics_service",
+    "replication_pull_service",
+    "replication_push_service",
+]
